test: extract management commands test to new file
