+++ aiarena/core/tests/tests.py
@@ -1,27 +1,18 @@
-import json
 import os
-from datetime import timedelta
-from io import StringIO
 
-from django.core import serializers
 from django.core.exceptions import ValidationError
 from django.core.files import File
-from django.core.management import CommandError, call_command
 from django.test import TestCase, TransactionTestCase
-from django.utils import timezone
 
 from constance import config
 
-from aiarena.core.api import Ladders, Matches
-from aiarena.core.management.commands import cleanupreplays
+from aiarena.core.api import Matches
 from aiarena.core.models import (
     Bot,
     Competition,
     CompetitionParticipation,
     Map,
     Match,
-    MatchParticipation,
-    Result,
     Round,
     User,
 )
@@ -379,300 +370,3 @@ class CompetitionsTestCase(FullDataSetMixin, TransactionTestCase):
         a_terran_bot = Bot.objects.filter(plays_race=terran).first()
         cp = CompetitionParticipation.objects.create(bot=a_terran_bot, competition=competition)
         cp.full_clean()  # causes validation to run
-
-
-class ManagementCommandTests(MatchReadyMixin, TransactionTestCase):
-    """
-    Tests for management commands
-    """
-
-    def test_cancel_matches(self):
-        # freeze competition2, so we can get anticipatable results
-        competition1 = Competition.objects.filter(status="open").first()
-        competition2 = Competition.objects.exclude(id=competition1.id).get()
-        competition2.freeze()
-
-        count = CompetitionParticipation.objects.filter(competition_id=competition1.id, active=True).count()
-        expectedMatchCountPerRound = int(count / 2 * (count - 1))
-
-        # test match doesn't exist
-        with self.assertRaisesMessage(CommandError, 'Match "12345" does not exist'):
-            call_command("cancelmatches", "12345")
-
-        # test match successfully cancelled
-        self.client.login(username="staff_user", password="x")
-        response = self._post_to_matches()
-        self.assertEqual(response.status_code, 201)
-        match_id = response.data["id"]
-
-        out = StringIO()
-        call_command("cancelmatches", match_id, stdout=out)
-        self.assertIn(f'Successfully marked match "{match_id}" with MatchCancelled', out.getvalue())
-
-        # test result already exists
-        with self.assertRaisesMessage(CommandError, f'A result already exists for match "{match_id}"'):
-            call_command("cancelmatches", match_id)
-
-        # test that cancelling the match marks it as started.
-        self.assertIsNotNone(Match.objects.get(id=match_id).started)
-
-        response = self._post_to_matches()
-        self.assertEqual(response.status_code, 201)
-        active_match_id = response.data["id"]
-        out = StringIO()
-        call_command("cancelmatches", "--active", stdout=out)
-        self.assertIn(f'Successfully marked match "{active_match_id}" with MatchCancelled', out.getvalue())
-
-        # cancel the rest of the matches.
-        # start at 2 because we already cancelled some above
-        for x in range(2, expectedMatchCountPerRound):
-            response = self._post_to_matches()
-            self.assertEqual(response.status_code, 201)
-            match_id = response.data["id"]
-
-            out = StringIO()
-            call_command("cancelmatches", match_id, stdout=out)
-            self.assertIn(f'Successfully marked match "{match_id}" with MatchCancelled', out.getvalue())
-
-            # test result already exists
-            with self.assertRaisesMessage(CommandError, f'A result already exists for match "{match_id}"'):
-                call_command("cancelmatches", match_id)
-
-            # test that cancelling the match marks it as started.
-            self.assertIsNotNone(Match.objects.get(id=match_id).started)
-
-        # check that the round was correctly marked as finished
-        round = Match.objects.get(id=match_id).round
-        self.assertTrue(round.complete)
-        self.assertIsNotNone(round.finished)
-
-    def test_cleanup_replays_and_logs(self):
-        NUM_MATCHES = 12
-        self.test_client.login(self.staffUser1)
-
-        # freeze competition2, so we can get anticipatable results
-        competition1 = Competition.objects.filter(status="open").first()
-        competition2 = Competition.objects.exclude(id=competition1.id).get()
-        competition2.freeze()
-
-        # generate some matches so we have replays to delete...
-        for x in range(NUM_MATCHES):  # 12 = two rounds
-            response = self._post_to_matches()
-            self.assertEqual(response.status_code, 201, f"{response.status_code} {response.data}")
-            match_id = response.data["id"]
-            self._post_to_results(match_id, "Player1Win")
-
-        # double check the replay and log files exist
-        results = Result.objects.filter(replay_file__isnull=False)
-        self.assertEqual(results.count(), NUM_MATCHES)
-        # after we ensure the arena client log count matches, we can safely just use the above results list
-        results_logs = Result.objects.filter(arenaclient_log__isnull=False)
-        self.assertEqual(results_logs.count(), NUM_MATCHES)
-        participants = MatchParticipation.objects.filter(match_log__isnull=False)
-        self.assertEqual(participants.count(), NUM_MATCHES * 2)
-
-        # set the created time so they'll be purged
-        results.update(created=timezone.now() - timedelta(days=cleanupreplays.Command._DEFAULT_DAYS_LOOKBACK + 1))
-
-        out = StringIO()
-        call_command("cleanupreplays", stdout=out)
-        self.assertIn(
-            "Cleaning up replays starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            f"{NUM_MATCHES} records gathered.\n"
-            f"Cleaned up {NUM_MATCHES} replays.",
-            out.getvalue(),
-        )
-
-        # ensure the job doesn't re-clean the same records when run again
-        out = StringIO()
-        call_command("cleanupreplays", stdout=out)
-        self.assertIn(
-            "Cleaning up replays starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            "0 records gathered.\n"
-            "Cleaned up 0 replays.",
-            out.getvalue(),
-        )
-
-        self.assertEqual(results.count(), NUM_MATCHES)
-        for result in results:
-            self.assertFalse(result.replay_file)
-
-        out = StringIO()
-        call_command("cleanuparenaclientlogfiles", stdout=out)
-        self.assertIn(
-            "Cleaning up arena client logfiles starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            f"{NUM_MATCHES} records gathered.\n"
-            f"Cleaned up {NUM_MATCHES} logfiles.",
-            out.getvalue(),
-        )
-
-        # ensure the job doesn't re-clean the same records when run again
-        out = StringIO()
-        call_command("cleanuparenaclientlogfiles", stdout=out)
-        self.assertIn(
-            "Cleaning up arena client logfiles starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            "0 records gathered.\n"
-            "Cleaned up 0 logfiles.",
-            out.getvalue(),
-        )
-
-        self.assertEqual(results.count(), NUM_MATCHES)
-        for result in results:
-            self.assertFalse(result.replay_file)
-
-        out = StringIO()
-        call_command("cleanupmatchlogfiles", stdout=out)
-        self.assertIn(
-            "Cleaning up match logfiles starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            f"{NUM_MATCHES * 2} records gathered.\n"
-            f"Cleaned up {NUM_MATCHES * 2} logfiles.",
-            out.getvalue(),
-        )
-
-        # ensure the job doesn't re-clean the same records when run again
-        out = StringIO()
-        call_command("cleanupmatchlogfiles", stdout=out)
-        self.assertIn(
-            "Cleaning up match logfiles starting from 30 days into the past...\n"
-            "Gathering records to clean...\n"
-            "0 records gathered.\n"
-            "Cleaned up 0 logfiles.",
-            out.getvalue(),
-        )
-
-        self.assertEqual(participants.count(), NUM_MATCHES * 2)
-        for participant in participants:
-            self.assertFalse(participant.match_log)
-
-    def test_generatestats(self):
-        self._generate_full_data_set()
-        out = StringIO()
-        call_command("generatestats", stdout=out)
-        self.assertIn("Done", out.getvalue())
-
-    def test_finalizecompetition(self):
-        def _sort_json(json_str: str):
-            json_list = json.loads(json_str)
-            return sorted(json_list, key=lambda x: x["pk"])
-
-        self._generate_full_data_set()
-
-        for competition in Competition.objects.all():
-            call_command("generatestats", "--competitionid", competition.id, "--finalize")
-
-        for competition in Competition.objects.all():
-            with self.assertRaisesMessage(
-                CommandError,
-                f"Competition {competition.id} is not closed! " "It must be closed before it can be finalized.",
-            ):
-                out = StringIO()
-                call_command("finalizecompetition", "--competitionid", competition.id, stdout=out)
-
-            # The front end uses get_competition_last_round_participants for displaying a closed competition's rankings
-            # so here we test that the displayed ranks don't change after finalization
-            pre_finalization_ranks = serializers.serialize(
-                "json", Ladders.get_competition_last_round_participants(competition)
-            )
-
-            # make sure it matches the known working method
-            pre_finalization_ranks_legacy = serializers.serialize(
-                "json", Ladders.get_competition_last_round_participants_legacy(competition)
-            )
-            self.assertEqual(_sort_json(pre_finalization_ranks), _sort_json(pre_finalization_ranks_legacy))
-
-            # fake competition closure
-            competition.status = "closed"
-            competition.save()
-
-            out = StringIO()
-            call_command("finalizecompetition", "--competitionid", competition.id, stdout=out)
-            self.assertIn(f"Competition {competition.id} has been finalized.", out.getvalue())
-
-            post_finalization_ranks = serializers.serialize(
-                "json", Ladders.get_competition_last_round_participants(competition)
-            )
-            self.assertEqual(pre_finalization_ranks, post_finalization_ranks)
-
-            # make sure it matches the known working method
-            post_finalization_ranks_legacy = serializers.serialize(
-                "json", Ladders.get_competition_last_round_participants_legacy(competition)
-            )
-            self.assertEqual(_sort_json(post_finalization_ranks), _sort_json(post_finalization_ranks_legacy))
-
-            out = StringIO()
-            call_command("finalizecompetition", "--competitionid", competition.id, stdout=out)
-            self.assertIn(f"WARNING: Competition {competition.id} is already finalized. Skipping...", out.getvalue())
-
-    def test_generatestats_competition(self):
-        self._generate_full_data_set()
-        out = StringIO()
-        for competition in Competition.objects.all():
-            call_command("generatestats", "--competitionid", competition.id, stdout=out)
-        self.assertIn("Done", out.getvalue())
-
-    def test_generatestats_competition_finalize(self):
-        self._generate_full_data_set()
-        for competition in Competition.objects.all():
-            out = StringIO()
-            call_command("generatestats", "--competitionid", competition.id, "--finalize", stdout=out)
-            self.assertIn("Done", out.getvalue())
-
-        # these should all fail
-        for competition in Competition.objects.all():
-            out = StringIO()
-            call_command("generatestats", "--competitionid", competition.id, stdout=out)
-            self.assertIn("WARNING: Skipping competition", out.getvalue())
-
-        for competition in Competition.objects.all():
-            out = StringIO()
-            call_command("generatestats", "--competitionid", competition.id, "--finalize", stdout=out)
-            self.assertIn("WARNING: Skipping competition", out.getvalue())
-
-    def test_generatestats_finalize_invalid(self):
-        with self.assertRaisesMessage(CommandError, "--finalize is only valid with --competitionid"):
-            call_command("generatestats", "--finalize")
-
-        with self.assertRaisesMessage(CommandError, "--finalize is only valid with --competitionid"):
-            call_command("generatestats", "--allcompetitions", "--finalize")
-
-    def test_generatestats_bot(self):
-        self._generate_full_data_set()
-        out = StringIO()
-        for bot in Bot.objects.all():
-            call_command("generatestats", "--botid", bot.id, stdout=out)
-        self.assertIn("Done", out.getvalue())
-
-    def test_seed(self):
-        out = StringIO()
-        call_command("seed", stdout=out)
-        self.assertIn('Done. User logins have a password of "x".', out.getvalue())
-
-    def test_check_bot_hashes(self):
-        call_command("checkbothashes")
-
-    def test_repair_bot_hashes(self):
-        call_command("repairbothashes")
-
-    def test_timeout_overtime_matches(self):
-        self.test_client.login(User.objects.get(username="arenaclient1"))
-
-        response = self.client.post("/api/arenaclient/matches/")
-        self.assertEqual(response.status_code, 201)
-
-        # save the match for modification
-        match1 = Match.objects.get(id=response.data["id"])
-
-        # set the created time back into the past long enough for it to cause a time out
-        match1.first_started = timezone.now() - (config.TIMEOUT_MATCHES_AFTER + timedelta(hours=1))
-        match1.save()
-
-        # this should trigger the bots to be forced out of the match
-        call_command("timeoutovertimematches")
-
-        # confirm a result was registered
-        self.assertTrue(match1.result is not None)
